move function only used in tests to tests
